chore: rename allreadyMinted to alreadyMinted, change description already-minted swagger
