+++ src/routes/token/alreadyMinted.ts
